requirements.txt
python-docx 
streamlit
pypdf
streamlit
python-dotenv
google-generativeai
ollama
